+++ packages/camera/camera_android/android/src/main/java/io/flutter/plugins/camera/CameraDeviceWrapper.java
@@ -4,6 +4,7 @@
 
 package io.flutter.plugins.camera;
 
+import android.annotation.SuppressLint;
 import android.annotation.TargetApi;
 import android.hardware.camera2.CameraAccessException;
 import android.hardware.camera2.CameraCaptureSession;
@@ -21,6 +22,7 @@ interface CameraDeviceWrapper {
   @NonNull
   CaptureRequest.Builder createCaptureRequest(int templateType) throws CameraAccessException;
 
+  @SuppressLint("UseRequiresApi")
   @TargetApi(Build.VERSION_CODES.P)
   void createCaptureSession(SessionConfiguration config) throws CameraAccessException;