+++ packages/camera/camera_android_camerax/lib/src/camerax_proxy.dart
@@ -76,7 +76,8 @@ class CameraXProxy {
   final void Function({
     BinaryMessenger? pigeonBinaryMessenger,
     PigeonInstanceManager? pigeonInstanceManager,
-  }) setUpGenericsProxy;
+  })
+  setUpGenericsProxy;
 
   /// Constructs [CameraSize].
   final CameraSize Function({
@@ -84,7 +85,8 @@ class CameraXProxy {
     required int height,
     BinaryMessenger? pigeon_binaryMessenger,
     PigeonInstanceManager? pigeon_instanceManager,
-  }) newCameraSize;
+  })
+  newCameraSize;
 
   /// Constructs [CameraIntegerRange].
   final CameraIntegerRange Function({
@@ -92,36 +94,41 @@ class CameraXProxy {
     required int upper,
     BinaryMessenger? pigeon_binaryMessenger,
     PigeonInstanceManager? pigeon_instanceManager,
-  }) newCameraIntegerRange;
+  })
+  newCameraIntegerRange;
 
   /// Constructs [Observer].
   final Observer<T> Function<T>({
     required void Function(Observer<T>, T) onChanged,
     BinaryMessenger? pigeon_binaryMessenger,
     PigeonInstanceManager? pigeon_instanceManager,
-  }) newObserver;
+  })
+  newObserver;
 
   /// Constructs [CameraSelector].
   final CameraSelector Function({
     LensFacing? requireLensFacing,
     BinaryMessenger? pigeon_binaryMessenger,
     PigeonInstanceManager? pigeon_instanceManager,
-  }) newCameraSelector;
+  })
+  newCameraSelector;
 
   /// Constructs [SystemServicesManager].
   final SystemServicesManager Function({
     required void Function(SystemServicesManager, String) onCameraError,
     BinaryMessenger? pigeon_binaryMessenger,
     PigeonInstanceManager? pigeon_instanceManager,
-  }) newSystemServicesManager;
+  })
+  newSystemServicesManager;
 
   /// Constructs [DeviceOrientationManager].
   final DeviceOrientationManager Function({
     required void Function(DeviceOrientationManager, String)
-        onDeviceOrientationChanged,
+    onDeviceOrientationChanged,
     BinaryMessenger? pigeon_binaryMessenger,
     PigeonInstanceManager? pigeon_instanceManager,
-  }) newDeviceOrientationManager;
+  })
+  newDeviceOrientationManager;
 
   /// Constructs [Preview].
   final Preview Function({
@@ -129,14 +136,16 @@ class CameraXProxy {
     ResolutionSelector? resolutionSelector,
     BinaryMessenger? pigeon_binaryMessenger,
     PigeonInstanceManager? pigeon_instanceManager,
-  }) newPreview;
+  })
+  newPreview;
 
   /// Constructs [VideoCapture].
   final VideoCapture Function({
     required VideoOutput videoOutput,
     BinaryMessenger? pigeon_binaryMessenger,
     PigeonInstanceManager? pigeon_instanceManager,
-  }) withOutputVideoCapture;
+  })
+  withOutputVideoCapture;
 
   /// Constructs [Recorder].
   final Recorder Function({
@@ -145,14 +154,16 @@ class CameraXProxy {
     QualitySelector? qualitySelector,
     BinaryMessenger? pigeon_binaryMessenger,
     PigeonInstanceManager? pigeon_instanceManager,
-  }) newRecorder;
+  })
+  newRecorder;
 
   /// Constructs [VideoRecordEventListener].
   final VideoRecordEventListener Function({
     required void Function(VideoRecordEventListener, VideoRecordEvent) onEvent,
     BinaryMessenger? pigeon_binaryMessenger,
     PigeonInstanceManager? pigeon_instanceManager,
-  }) newVideoRecordEventListener;
+  })
+  newVideoRecordEventListener;
 
   /// Constructs [ImageCapture].
   final ImageCapture Function({
@@ -161,7 +172,8 @@ class CameraXProxy {
     ResolutionSelector? resolutionSelector,
     BinaryMessenger? pigeon_binaryMessenger,
     PigeonInstanceManager? pigeon_instanceManager,
-  }) newImageCapture;
+  })
+  newImageCapture;
 
   /// Constructs [ResolutionStrategy].
   final ResolutionStrategy Function({
@@ -169,7 +181,8 @@ class CameraXProxy {
     required ResolutionStrategyFallbackRule fallbackRule,
     BinaryMessenger? pigeon_binaryMessenger,
     PigeonInstanceManager? pigeon_instanceManager,
-  }) newResolutionStrategy;
+  })
+  newResolutionStrategy;
 
   /// Constructs [ResolutionSelector].
   final ResolutionSelector Function({
@@ -178,7 +191,8 @@ class CameraXProxy {
     ResolutionFilter? resolutionFilter,
     BinaryMessenger? pigeon_binaryMessenger,
     PigeonInstanceManager? pigeon_instanceManager,
-  }) newResolutionSelector;
+  })
+  newResolutionSelector;
 
   /// Constructs [AspectRatioStrategy].
   final AspectRatioStrategy Function({
@@ -186,7 +200,8 @@ class CameraXProxy {
     required AspectRatioStrategyFallbackRule fallbackRule,
     BinaryMessenger? pigeon_binaryMessenger,
     PigeonInstanceManager? pigeon_instanceManager,
-  }) newAspectRatioStrategy;
+  })
+  newAspectRatioStrategy;
 
   /// Constructs [ImageAnalysis].
   final ImageAnalysis Function({
@@ -194,14 +209,16 @@ class CameraXProxy {
     ResolutionSelector? resolutionSelector,
     BinaryMessenger? pigeon_binaryMessenger,
     PigeonInstanceManager? pigeon_instanceManager,
-  }) newImageAnalysis;
+  })
+  newImageAnalysis;
 
   /// Constructs [Analyzer].
   final Analyzer Function({
     required void Function(Analyzer, ImageProxy) analyze,
     BinaryMessenger? pigeon_binaryMessenger,
     PigeonInstanceManager? pigeon_instanceManager,
-  }) newAnalyzer;
+  })
+  newAnalyzer;
 
   /// Constructs [QualitySelector].
   final QualitySelector Function({
@@ -209,7 +226,8 @@ class CameraXProxy {
     FallbackStrategy? fallbackStrategy,
     BinaryMessenger? pigeon_binaryMessenger,
     PigeonInstanceManager? pigeon_instanceManager,
-  }) fromQualitySelector;
+  })
+  fromQualitySelector;
 
   /// Constructs [QualitySelector].
   final QualitySelector Function({
@@ -217,42 +235,48 @@ class CameraXProxy {
     FallbackStrategy? fallbackStrategy,
     BinaryMessenger? pigeon_binaryMessenger,
     PigeonInstanceManager? pigeon_instanceManager,
-  }) fromOrderedListQualitySelector;
+  })
+  fromOrderedListQualitySelector;
 
   /// Constructs [FallbackStrategy].
   final FallbackStrategy Function({
     required VideoQuality quality,
     BinaryMessenger? pigeon_binaryMessenger,
     PigeonInstanceManager? pigeon_instanceManager,
-  }) higherQualityOrLowerThanFallbackStrategy;
+  })
+  higherQualityOrLowerThanFallbackStrategy;
 
   /// Constructs [FallbackStrategy].
   final FallbackStrategy Function({
     required VideoQuality quality,
     BinaryMessenger? pigeon_binaryMessenger,
     PigeonInstanceManager? pigeon_instanceManager,
-  }) higherQualityThanFallbackStrategy;
+  })
+  higherQualityThanFallbackStrategy;
 
   /// Constructs [FallbackStrategy].
   final FallbackStrategy Function({
     required VideoQuality quality,
     BinaryMessenger? pigeon_binaryMessenger,
     PigeonInstanceManager? pigeon_instanceManager,
-  }) lowerQualityOrHigherThanFallbackStrategy;
+  })
+  lowerQualityOrHigherThanFallbackStrategy;
 
   /// Constructs [FallbackStrategy].
   final FallbackStrategy Function({
     required VideoQuality quality,
     BinaryMessenger? pigeon_binaryMessenger,
     PigeonInstanceManager? pigeon_instanceManager,
-  }) lowerQualityThanFallbackStrategy;
+  })
+  lowerQualityThanFallbackStrategy;
 
   /// Constructs [FocusMeteringActionBuilder].
   final FocusMeteringActionBuilder Function({
     required MeteringPoint point,
     BinaryMessenger? pigeon_binaryMessenger,
     PigeonInstanceManager? pigeon_instanceManager,
-  }) newFocusMeteringActionBuilder;
+  })
+  newFocusMeteringActionBuilder;
 
   /// Constructs [FocusMeteringActionBuilder].
   FocusMeteringActionBuilder Function({
@@ -260,35 +284,40 @@ class CameraXProxy {
     required MeteringMode mode,
     BinaryMessenger? pigeon_binaryMessenger,
     PigeonInstanceManager? pigeon_instanceManager,
-  }) withModeFocusMeteringActionBuilder;
+  })
+  withModeFocusMeteringActionBuilder;
 
   /// Constructs [CaptureRequestOptions].
   CaptureRequestOptions Function({
     required Map<CaptureRequestKey, Object?> options,
     BinaryMessenger? pigeon_binaryMessenger,
     PigeonInstanceManager? pigeon_instanceManager,
-  }) newCaptureRequestOptions;
+  })
+  newCaptureRequestOptions;
 
   /// Constructs [Camera2CameraControl].
   Camera2CameraControl Function({
     required CameraControl cameraControl,
     BinaryMessenger? pigeon_binaryMessenger,
     PigeonInstanceManager? pigeon_instanceManager,
-  }) fromCamera2CameraControl;
+  })
+  fromCamera2CameraControl;
 
   /// Constructs [ResolutionFilter].
   final ResolutionFilter Function({
     required CameraSize preferredSize,
     BinaryMessenger? pigeon_binaryMessenger,
     PigeonInstanceManager? pigeon_instanceManager,
-  }) createWithOnePreferredSizeResolutionFilter;
+  })
+  createWithOnePreferredSizeResolutionFilter;
 
   /// Constructs [Camera2CameraInfo].
   final Camera2CameraInfo Function({
     required CameraInfo cameraInfo,
     BinaryMessenger? pigeon_binaryMessenger,
     PigeonInstanceManager? pigeon_instanceManager,
-  }) fromCamera2CameraInfo;
+  })
+  fromCamera2CameraInfo;
 
   /// Constructs [DisplayOrientedMeteringPointFactory].
   DisplayOrientedMeteringPointFactory Function({
@@ -297,13 +326,15 @@ class CameraXProxy {
     required double height,
     BinaryMessenger? pigeon_binaryMessenger,
     PigeonInstanceManager? pigeon_instanceManager,
-  }) newDisplayOrientedMeteringPointFactory;
+  })
+  newDisplayOrientedMeteringPointFactory;
 
   /// Calls to [ProcessCameraProvider.getInstance].
   final Future<ProcessCameraProvider> Function({
     BinaryMessenger? pigeon_binaryMessenger,
     PigeonInstanceManager? pigeon_instanceManager,
-  }) getInstanceProcessCameraProvider;
+  })
+  getInstanceProcessCameraProvider;
 
   /// Calls to [QualitySelector.getResolution].
   final Future<CameraSize?> Function(
@@ -311,7 +342,8 @@ class CameraXProxy {
     VideoQuality, {
     BinaryMessenger? pigeon_binaryMessenger,
     PigeonInstanceManager? pigeon_instanceManager,
-  }) getResolutionQualitySelector;
+  })
+  getResolutionQualitySelector;
 
   /// Calls to [CameraSelector.defaultBackCamera].
   final CameraSelector Function() defaultBackCameraCameraSelector;
@@ -321,26 +353,26 @@ class CameraXProxy {
 
   /// Calls to [ResolutionStrategy.highestAvailableStrategy].
   final ResolutionStrategy Function()
-      highestAvailableStrategyResolutionStrategy;
+  highestAvailableStrategyResolutionStrategy;
 
   /// Calls to [AspectRatioStrategy.ratio_16_9FallbackAutoStrategy].
   final AspectRatioStrategy Function()
-      ratio_16_9FallbackAutoStrategyAspectRatioStrategy;
+  ratio_16_9FallbackAutoStrategyAspectRatioStrategy;
 
   /// Calls to [AspectRatioStrategy.ratio_4_3FallbackAutoStrategy].
   final AspectRatioStrategy Function()
-      ratio_4_3FallbackAutoStrategyAspectRatioStrategy;
+  ratio_4_3FallbackAutoStrategyAspectRatioStrategy;
 
   /// Calls to [CaptureRequest.controlAELock].
   CaptureRequestKey Function() controlAELockCaptureRequest;
 
   /// Calls to [CameraCharacteristics.infoSupportedHardwareLevel].
   final CameraCharacteristicsKey Function()
-      infoSupportedHardwareLevelCameraCharacteristics;
+  infoSupportedHardwareLevelCameraCharacteristics;
 
   /// Calls to [CameraCharacteristics.sensorOrientation].
   final CameraCharacteristicsKey Function()
-      sensorOrientationCameraCharacteristics;
+  sensorOrientationCameraCharacteristics;
 
   static CameraSelector _defaultBackCameraCameraSelector() =>
       CameraSelector.defaultBackCamera;
@@ -352,19 +384,19 @@ class CameraXProxy {
       ResolutionStrategy.highestAvailableStrategy;
 
   static AspectRatioStrategy
-      _ratio_16_9FallbackAutoStrategyAspectRatioStrategy() =>
-          AspectRatioStrategy.ratio_16_9FallbackAutoStrategy;
+  _ratio_16_9FallbackAutoStrategyAspectRatioStrategy() =>
+      AspectRatioStrategy.ratio_16_9FallbackAutoStrategy;
 
   static AspectRatioStrategy
-      _ratio_4_3FallbackAutoStrategyAspectRatioStrategy() =>
-          AspectRatioStrategy.ratio_4_3FallbackAutoStrategy;
+  _ratio_4_3FallbackAutoStrategyAspectRatioStrategy() =>
+      AspectRatioStrategy.ratio_4_3FallbackAutoStrategy;
 
   static CaptureRequestKey _controlAELockCaptureRequest() =>
       CaptureRequest.controlAELock;
 
   static CameraCharacteristicsKey
-      _infoSupportedHardwareLevelCameraCharacteristics() =>
-          CameraCharacteristics.infoSupportedHardwareLevel;
+  _infoSupportedHardwareLevelCameraCharacteristics() =>
+      CameraCharacteristics.infoSupportedHardwareLevel;
 
   static CameraCharacteristicsKey _sensorOrientationCameraCharacteristics() =>
       CameraCharacteristics.sensorOrientation;