+++ packages/camera/camera_android_camerax/pigeons/camerax_library.dart
@@ -18,7 +18,6 @@ import 'package:pigeon/pigeon.dart';
     ),
   ),
 )
-
 /// Immutable class for describing width and height dimensions in pixels.
 ///
 /// See https://developer.android.com/reference/android/util/Size.html.
@@ -518,6 +517,9 @@ abstract class VideoRecordEventListener {
   ),
 )
 abstract class PendingRecording {
+  /// Enables/disables audio to be recorded for this recording.
+  PendingRecording withAudioEnabled(bool initialMuted);
+
   /// Starts the recording, making it an active recording.
   Recording start(VideoRecordEventListener listener);
 }