+++ packages/quick_actions/quick_actions_android/android/src/main/java/io/flutter/plugins/quickactions/QuickActions.java
@@ -4,6 +4,7 @@
 
 package io.flutter.plugins.quickactions;
 
+import android.annotation.SuppressLint;
 import android.annotation.TargetApi;
 import android.app.Activity;
 import android.content.Context;
@@ -121,6 +122,7 @@ final class QuickActions implements AndroidQuickActionsApi {
     return launchAction;
   }
 
+  @SuppressLint("UseRequiresApi")
   @TargetApi(Build.VERSION_CODES.N_MR1)
   private List<ShortcutInfoCompat> shortcutItemMessageToShortcutInfo(
       @NonNull List<ShortcutItemMessage> shortcuts) {