+++ packages/video_player/video_player_android/lib/src/android_video_player.dart
@@ -84,8 +84,9 @@ class AndroidVideoPlayer extends VideoPlayerPlatform {
     final int playerId = await _api.create(message);
     _playerViewStates[playerId] = switch (options.viewType) {
       // playerId is also the textureId when using texture view.
-      VideoViewType.textureView =>
-        _VideoPlayerTextureViewState(textureId: playerId),
+      VideoViewType.textureView => _VideoPlayerTextureViewState(
+        textureId: playerId,
+      ),
       VideoViewType.platformView => const _VideoPlayerPlatformViewState(),
     };
 
@@ -132,23 +133,23 @@ class AndroidVideoPlayer extends VideoPlayerPlatform {
 
   @override
   Stream<VideoEvent> videoEventsFor(int playerId) {
-    return _eventChannelFor(playerId)
-        .receiveBroadcastStream()
-        .map((dynamic event) {
+    return _eventChannelFor(playerId).receiveBroadcastStream().map((
+      dynamic event,
+    ) {
       final Map<dynamic, dynamic> map = event as Map<dynamic, dynamic>;
       switch (map['event']) {
         case 'initialized':
           return VideoEvent(
             eventType: VideoEventType.initialized,
             duration: Duration(milliseconds: map['duration'] as int),
-            size: Size((map['width'] as num?)?.toDouble() ?? 0.0,
-                (map['height'] as num?)?.toDouble() ?? 0.0),
+            size: Size(
+              (map['width'] as num?)?.toDouble() ?? 0.0,
+              (map['height'] as num?)?.toDouble() ?? 0.0,
+            ),
             rotationCorrection: map['rotationCorrection'] as int? ?? 0,
           );
         case 'completed':
-          return VideoEvent(
-            eventType: VideoEventType.completed,
-          );
+          return VideoEvent(eventType: VideoEventType.completed);
         case 'bufferingUpdate':
           final List<dynamic> values = map['values'] as List<dynamic>;
 
@@ -173,9 +174,7 @@ class AndroidVideoPlayer extends VideoPlayerPlatform {
 
   @override
   Widget buildView(int playerId) {
-    return buildViewWithOptions(
-      VideoViewOptions(playerId: playerId),
-    );
+    return buildViewWithOptions(VideoViewOptions(playerId: playerId));
   }
 
   @override
@@ -184,10 +183,12 @@ class AndroidVideoPlayer extends VideoPlayerPlatform {
     final _VideoPlayerViewState? viewState = _playerViewStates[playerId];
 
     return switch (viewState) {
-      _VideoPlayerTextureViewState(:final int textureId) =>
-        Texture(textureId: textureId),
+      _VideoPlayerTextureViewState(:final int textureId) => Texture(
+        textureId: textureId,
+      ),
       _VideoPlayerPlatformViewState() => PlatformViewPlayer(playerId: playerId),
-      null => throw Exception(
+      null =>
+        throw Exception(
           'Could not find corresponding view type for playerId: $playerId',
         ),
     };
@@ -204,11 +205,11 @@ class AndroidVideoPlayer extends VideoPlayerPlatform {
 
   static const Map<VideoFormat, String> _videoFormatStringMap =
       <VideoFormat, String>{
-    VideoFormat.ss: 'ss',
-    VideoFormat.hls: 'hls',
-    VideoFormat.dash: 'dash',
-    VideoFormat.other: 'other',
-  };
+        VideoFormat.ss: 'ss',
+        VideoFormat.hls: 'hls',
+        VideoFormat.dash: 'dash',
+        VideoFormat.other: 'other',
+      };
 
   DurationRange _toDurationRange(dynamic value) {
     final List<dynamic> pair = value as List<dynamic>;
@@ -237,9 +238,7 @@ sealed class _VideoPlayerViewState {
 /// Represents the state of a video player view that uses a texture.
 final class _VideoPlayerTextureViewState extends _VideoPlayerViewState {
   /// Creates a new instance of [_VideoPlayerTextureViewState].
-  const _VideoPlayerTextureViewState({
-    required this.textureId,
-  });
+  const _VideoPlayerTextureViewState({required this.textureId});
 
   /// The ID of the texture used by the video player.
   final int textureId;