+++ packages/webview_flutter/webview_flutter_wkwebview/darwin/webview_flutter_wkwebview/Sources/webview_flutter_wkwebview/AuthenticationChallengeResponseProxyAPIDelegate.swift
@@ -34,6 +34,17 @@ class AuthenticationChallengeResponseProxyAPIDelegate:
     return AuthenticationChallengeResponse(disposition: nativeDisposition, credential: credential)
   }
 
+  func createAsync(
+    pigeonApi: PigeonApiAuthenticationChallengeResponse,
+    disposition: UrlSessionAuthChallengeDisposition, credential: URLCredential?,
+    completion: @escaping (Result<AuthenticationChallengeResponse, Error>) -> Void
+  ) {
+    completion(
+      Result.success(
+        try! pigeonDefaultConstructor(
+          pigeonApi: pigeonApi, disposition: disposition, credential: credential)))
+  }
+
   func disposition(
     pigeonApi: PigeonApiAuthenticationChallengeResponse,
     pigeonInstance: AuthenticationChallengeResponse